requirements.txt
flask
flask-cors
numpy
matplotlib
PIL
imageio
imageio-ffmpeg
python-dotenv
